final rails omnicalc

+++ app/controllers/square_results_controller.rb
@@ -1,6 +1,6 @@
 class SquareResultsController < ApplicationController
   def show_square_results
-    @usernumber = params.fetch("usernumber").to_i
+    @usernumber = params.fetch("usernumber").to_f
     @squared_user_number = @usernumber ** 2
     render({:template => "pages/square_results" })
   end
